renamed progress-indicator.ts to busy-indicator.ts and removed useless tests
