+++ src/codeview/busy-indicator.ts
